requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tensorflow==2.13.0
yfinance==0.2.28
python-dotenv==1.0.0